Disable cabal testing using GHC 8.8.3 and 8.10.1
The ginger library has dependency issues.  When using older versions of
GHC, older versions of ginger compile while newer versions do not.  When
using newer versions of GHC, newer versions of ginger compile while
older versions do not.  Unfortunately, there is no way to specify which
version of the library to use for each version of GHC in the cabal file,
making it impossible for this program to support using both older and
newer versions of GHC with a single cabal file.

Note that the ginger version to use for each GHC version is specified
when using Stack, either explicitly in stack.yaml or implicitly in the
Stack package set.  The program builds successfully in all of the tested
GHC versions (including 8.10.1) using Stack.